requirements.txt
azure-functions
azure-storage-blob
requests
beautifulsoup4
feedparser
openai